+++ CHANGES.md
@@ -3,6 +3,12 @@
 This document describes the relevant changes between releases of the OCM API
 SDK.
 
+## 0.1.261 Apr 14 2022
+
+- Update to model 0.0.193:
+  - Add limitedSupportReasonCount to cluster status struct.
+  - Add inflight check API.
+
 ## 0.1.260 Apr 11 2022
 
 - Update to model 0.0.191: